// app/redux/plan/types.ts
export enum PLANS_ACTION_TYPES {
  CREATE_ASSET_CLASS_PLAN_ERROR = 'plan/CREATE_ASSET_CLASS_PLAN_ERROR',

  CREATE_ASSET_CLASS_PLAN_REQUEST = 'plan/CREATE_ASSET_CLASS_PLAN_REQUEST',
  CREATE_ASSET_CLASS_PLAN_SUCCESS = 'plan/CREATE_ASSET_CLASS_PLAN_SUCCESS',
  CREATE_BUILD_WEALTH_PLAN_ERROR = 'plan/CREATE_BUILD_WEALTH_PLAN_ERROR',

  CREATE_BUILD_WEALTH_PLAN_REQUEST = 'plan/CREATE_BUILD_WEALTH_PLAN_REQUEST',
  CREATE_BUILD_WEALTH_PLAN_SUCCESS = 'plan/CREATE_BUILD_WEALTH_PLAN_SUCCESS',
  CREATE_GIFTED_PLAN_ERROR = 'plan/CREATE_GIFTED_PLAN_ERROR',

  CREATE_GIFTED_PLAN_REQUEST = 'plan/CREATE_GIFTED_PLAN_REQUEST',
  CREATE_GIFTED_PLAN_SUCCESS = 'plan/CREATE_GIFTED_PLAN_SUCCESS',
  CREATE_INVESTMENT_PROFILE_ERROR = 'plan/CREATE_INVESTMENT_PROFILE_ERROR',

  CREATE_INVESTMENT_PROFILE_REQUEST = 'plan/CREATE_INVESTMENT_PROFILE_REQUEST',
  CREATE_INVESTMENT_PROFILE_SUCCESS = 'plan/CREATE_INVESTMENT_PROFILE_SUCCESS',
  DELETE_PLAN_ERROR = 'plan/DELETE_PLAN_ERROR',

  DELETE_PLAN_REQUEST = 'plan/DELETE_PLAN_REQUEST',
  DELETE_PLAN_SUCCESS = 'plan/DELETE_PLAN_SUCCESS',
  EDIT_PLAN_ERROR = 'plan/EDIT_PLAN_ERROR',

  EDIT_PLAN_REQUEST = 'plan/EDIT_PLAN_REQUEST',
  EDIT_PLAN_SUCCESS = 'plan/EDIT_PLAN_SUCCESS',
  FETCH_BUILD_WEALTH_PLAN_ERROR = 'plan/FETCH_BUILD_WEALTH_PLAN_ERROR',

  FETCH_BUILD_WEALTH_PLAN_REQUEST = 'plan/FETCH_BUILD_WEALTH_PLAN_REQUEST',
  FETCH_BUILD_WEALTH_PLAN_SUCCESS = 'plan/FETCH_BUILD_WEALTH_PLAN_SUCCESS',
  FETCH_GIFTED_PLAN_ERROR = 'plan/FETCH_GIFTED_PLAN_ERROR',

  FETCH_GIFTED_PLAN_REQUEST = 'plan/FETCH_GIFTED_PLAN_REQUEST',
  FETCH_GIFTED_PLAN_SUCCESS = 'plan/FETCH_GIFTED_PLAN_SUCCESS',
  FETCH_GOAL_PLANS_ERROR = 'plan/FETCH_GOAL_PLANS_ERROR',

  FETCH_GOAL_PLANS_REQUEST = 'plan/FETCH_GOAL_PLANS_REQUEST',
  FETCH_GOAL_PLANS_SUCCESS = 'plan/FETCH_GOAL_PLANS_SUCCESS',
  FETCH_INVESTMENT_PROFILE_ERROR = 'plan/FETCH_INVESTMENT_PROFILE_ERROR',

  FETCH_INVESTMENT_PROFILE_REQUEST = 'plan/FETCH_INVESTMENT_PROFILE_REQUEST',
  FETCH_INVESTMENT_PROFILE_SUCCESS = 'plan/FETCH_INVESTMENT_PROFILE_SUCCESS',
  FETCH_PLAN_HISTORY_ERROR = 'plan/FETCH_PLAN_HISTORY_ERROR',

  FETCH_PLAN_HISTORY_REQUEST = 'plan/FETCH_PLAN_HISTORY_REQUEST',
  FETCH_PLAN_HISTORY_SUCCESS = 'plan/FETCH_PLAN_HISTORY_SUCCESS',
  FETCH_USER_PLANS_ERROR = 'plan/FETCH_USER_PLANS_ERROR',

  FETCH_USER_PLANS_REQUEST = 'plan/FETCH_USER_PLANS_REQUEST',
  FETCH_USER_PLANS_SUCCESS = 'plan/FETCH_USER_PLANS_SUCCESS',

  FETCH_USER_PLAN_ERROR = 'plan/FETCH_USER_PLAN_ERROR',
  FETCH_USER_PLAN_REQUEST = 'plan/FETCH_USER_PLAN_REQUEST',
  FETCH_USER_PLAN_SUCCESS = 'plan/FETCH_USER_PLAN_SUCCESS',

  FUND_PLAN_WITH_WALLET_ERROR = 'plan/FUND_PLAN_WITH_WALLET_ERROR',
  FUND_PLAN_WITH_WALLET_REQUEST = 'plan/FUND_PLAN_WITH_WALLET_REQUEST',
  FUND_PLAN_WITH_WALLET_SUCCESS = 'plan/FUND_PLAN_WITH_WALLET_SUCCESS',

  GET_ASSET_CLASSES_ERROR = 'plan/GET_ASSET_CLASSES_ERROR',
  GET_ASSET_CLASSES_REQUEST = 'plan/GET_ASSET_CLASSES_REQUEST',
  GET_ASSET_CLASSES_SUCCESS = 'plan/GET_ASSET_CLASSES_SUCCESS',

  GET_PLAN_PROJECTION_ERROR = 'plan/GET_PLAN_PROJECTION_ERROR',
  GET_PLAN_PROJECTION_REQUEST = 'plan/GET_PLAN_PROJECTION_REQUEST',
  GET_PLAN_PROJECTION_SUCCESS = 'plan/GET_PLAN_PROJECTION_SUCCESS',

  GET_RISE_REAL_ESTATE_PORTFOLIO_ASSETS = 'plan/GET_RISE_REAL_ESTATE_PORTFOLIO_ASSETS',
  GET_RISE_STOCK_PORTFOLIO_ASSETS = 'plan/GET_RISE_STOCK_PORTFOLIO_ASSETS',
  JOIN_CHALLENGE_ERROR = 'challenge/JOIN_CHALLENGE_ERROR',

  JOIN_CHALLENGE_REQUEST = 'challenge/JOIN_CHALLENGE_REQUEST',
  JOIN_CHALLENGE_SUCCESS = 'challenge/JOIN_CHALLENGE_SUCCESS',
  REINVEST_PLAN_ERROR = 'plan/REINVEST_PLAN_ERROR',

  REINVEST_PLAN_REQUEST = 'plan/REINVEST_PLAN_REQUEST',
  REINVEST_PLAN_SUCCESS = 'plan/REINVEST_PLAN_SUCCESS',
  RESET_PLAN_DATA = 'plan/RESET_PLAN_DATA',

  RESET_PLAN_TEMP_DATA = 'plan/RESET_PLAN_TEMP_DATA',
  SAVE_PLAN_TEMP_DATA = 'plan/SAVE_PLAN_TEMP_DATA',
  SET_ASSET_CLASS_SUGGESTION = 'plan/SET_ASSET_CLASS_SUGGESTION',

  SET_INVESTMENT_DATA = 'plan/SET_INVESTMENT_DATA',
  UPDATE_GIFTED_PLAN_ERROR = 'plan/UPDATE_GIFTED_PLAN_ERROR',
  UPDATE_GIFTED_PLAN_REQUEST = 'plan/UPDATE_GIFTED_PLAN_REQUEST',

  UPDATE_GIFTED_PLAN_SUCCESS = 'plan/UPDATE_GIFTED_PLAN_SUCCESS',
  UPDATE_PLAN_AMOUNT_SUCCESS = 'plan/UPDATE_PLAN_AMOUNT_SUCCESS',

  UPLOAD_PLAN_IMAGE_ERROR = 'plan/UPLOAD_PLAN_IMAGE_ERROR',
  UPLOAD_PLAN_IMAGE_REQUEST = 'plan/UPLOAD_PLAN_IMAGE_REQUEST',
  UPLOAD_PLAN_IMAGE_SUCCESS = 'plan/UPLOAD_PLAN_IMAGE_SUCCESS',
  WATCH_PLAN_CHALLENGE_UPDATE = 'plan/WATCH_PLAN_CHALLENGE_UPDATE'
}

export type SetInvestmentProfile = {
  investmentProfile: any
  type: PLANS_ACTION_TYPES.CREATE_INVESTMENT_PROFILE_SUCCESS
}

export type FetchInvestmentProfileSuccess = {
  investmentProfile: any
  type: PLANS_ACTION_TYPES.FETCH_INVESTMENT_PROFILE_SUCCESS
}

export type SetInvestmentData = {
  investmentData: any
  type: PLANS_ACTION_TYPES.SET_INVESTMENT_DATA
}

export type SetAssetClassSuggestion = {
  assetClasses: any
  type: PLANS_ACTION_TYPES.SET_ASSET_CLASS_SUGGESTION
}

export type SetBuildWealthPlanSuccess = {
  buildWealthPlan: any
  type: PLANS_ACTION_TYPES.CREATE_BUILD_WEALTH_PLAN_SUCCESS
}

export type SetAssetClassPlanSuccess = {
  assetClassPlan: any
  type: PLANS_ACTION_TYPES.CREATE_ASSET_CLASS_PLAN_SUCCESS
}

export type UploadPlanImageSuccess = {
  data: {
    planId: number
    url: string
  }
  type: PLANS_ACTION_TYPES.UPLOAD_PLAN_IMAGE_SUCCESS
}

export type FetchUserPlanSuccess = {
  plan: IPlan
  type: PLANS_ACTION_TYPES.FETCH_USER_PLAN_SUCCESS
}

export type FetchBuildWealthPlanSuccess = {
  buildWealthPlan: IPlan
  type: PLANS_ACTION_TYPES.FETCH_BUILD_WEALTH_PLAN_SUCCESS
}

export type FundUserPlanSuccess = {
  amount: string
  type: PLANS_ACTION_TYPES.FUND_PLAN_WITH_WALLET_SUCCESS
  userPlanId: number
}

export type FetchUserPlansSuccess = {
  type: PLANS_ACTION_TYPES.FETCH_USER_PLANS_SUCCESS
  userPlans: {
    data: IPlan[]
    meta: {
      limit: number
      offset: number
      total: number
    }
  }
}

export type SavePlanTempData = {
  tempPlanData: any
  type: PLANS_ACTION_TYPES.SAVE_PLAN_TEMP_DATA
}

export type GetPlanProjectionSuccess = {
  projection: PlanProjection
  type: PLANS_ACTION_TYPES.GET_PLAN_PROJECTION_SUCCESS
}

export type EditPlanSuccess = {
  plan: IPlan
  type: PLANS_ACTION_TYPES.EDIT_PLAN_SUCCESS
}

export type GetAssetClassesSuccess = {
  assetClasses: any
  type: PLANS_ACTION_TYPES.GET_ASSET_CLASSES_SUCCESS
}

export type DeletePlanSuccess = {
  planId: number
  type: PLANS_ACTION_TYPES.DELETE_PLAN_SUCCESS
}

export type UpdatePlanAmountSuccess = {
  amount: string
  hasMatured: boolean
  planId: number
  type: PLANS_ACTION_TYPES.UPDATE_PLAN_AMOUNT_SUCCESS
  updateType: 'credit-plan' | 'debit-plan'
}

export type FetchGiftedPlanSuccess = {
  giftedPlan: GiftPlanData
  type: PLANS_ACTION_TYPES.CREATE_GIFTED_PLAN_SUCCESS
}

export type UpdateGiftedPlanSuccess = {
  plan: IPlan
  status: string
  type: PLANS_ACTION_TYPES.UPDATE_GIFTED_PLAN_SUCCESS
}

export type JoinChallengeSuccess = {
  plan: IPlan
  type: PLANS_ACTION_TYPES.JOIN_CHALLENGE_SUCCESS
}

export type WatchPlanChallengeSuccess = {
  plan: IPlan
  type: PLANS_ACTION_TYPES.WATCH_PLAN_CHALLENGE_UPDATE
}

export type FetchPlanHistorySuccess = {
  planHistory: IPlan[]
  planId: number
  type: PLANS_ACTION_TYPES.FETCH_PLAN_HISTORY_SUCCESS
}

export type FetchGoalPlansSuccess = {
  goalPlans: GoalPlanData[]
  type: PLANS_ACTION_TYPES.FETCH_GOAL_PLANS_SUCCESS
}

export type UnmatchedAction = {
  type: 'UNMATCHED_ACTION'
}

export type PlansActions =
  | SetInvestmentProfile
  | SetBuildWealthPlanSuccess
  | FetchUserPlanSuccess
  | SavePlanTempData
  | GetPlanProjectionSuccess
  | UploadPlanImageSuccess
  | GetAssetClassesSuccess
  | SetAssetClassPlanSuccess
  | FetchBuildWealthPlanSuccess
  | EditPlanSuccess
  | DeletePlanSuccess
  | UpdatePlanAmountSuccess
  | FetchGiftedPlanSuccess
  | UpdateGiftedPlanSuccess
  | JoinChallengeSuccess
  | WatchPlanChallengeSuccess
  | FetchPlanHistorySuccess
  | FetchGoalPlansSuccess
  | UnmatchedAction